Разделил компоненты Chat, Chats и Profile на презентационные и контейнеры
