+++ src/components/Profile/Profile.js
@@ -1,12 +1,9 @@
-import '../css/style.css';
+import '../../css/style.css';
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import MessageForm from './MessageForm';
+import MessageForm from '../MessageForm/MessageForm';
 import { withStyles } from '@material-ui/core/styles';
 import Checkbox from '@material-ui/core/Checkbox';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
-import { changeIsOnline, changeName } from '../actions/profile';
-import { profileSelector } from '../selectors/profile';
 
 const GreenCheckbox = withStyles({
 	root: {
@@ -18,37 +15,37 @@ const GreenCheckbox = withStyles({
 	checked: {},
 })((props) => <Checkbox color="default" {...props} />);
 
-export default function Profile() {
-	const { age, name, isOnline } = useSelector(profileSelector);
-
-	const dispatch = useDispatch();
-
-	const handleChangeIsOnline = (event) => {
-		dispatch(changeIsOnline(event.target.checked))
-	}
-
-	const handleChangeName = (newUserName) => {
-		dispatch(changeName(newUserName));
-	}
+export default function Profile(props) {
+	const {
+		age,
+		name,
+		isOnline,
+		handleChangeIsOnline,
+		handleChangeName
+	} = props;
 
 	return (
 		<main className="app__main main">
 			<section className="profile container">
 				<h2 className="profile__title visually-hidden">Profile page</h2>
-				<div className="profile__info">
-					<p className="profile__name">Name: {name}</p>
-					<p className="profile__age">Age: {age}</p>
-					<FormControlLabel
-						className="profile__online-status"
-						control={<GreenCheckbox checked={isOnline} onChange={handleChangeIsOnline} name="checkedG" />}
-						label="Is user online?"
-					/>
+				<div className="profile__info info">
+					<p className="info__name">Name: {name}</p>
+					<p className="info__age">Age: {age}</p>
 				</div>
-				<div className="chats__input">
+				<div className="profile__settings settings">
+					<FormControlLabel
+						className="settings__online-status"
+						control={<GreenCheckbox checked={isOnline} onChange={handleChangeIsOnline} name="checkedG" />}
+						label='Показывать статус "в сети"?'
+					/>
 					<MessageForm
+						className="settings__name"
 						label="Введите новое имя..."
 						onSubmit={handleChangeName}
 					/>
+				</div>
+				<div className="chats__input">
+
 				</div>
 			</section>
 		</main>
